add Trấnctional for service and fix math.IntToExtra

+++ Backend/src/main/java/com/hospital/hospitalmis/repository/ImagingProcedureRepository.java
@@ -3,4 +3,4 @@ package com.hospital.hospitalmis.repository;
 import com.hospital.hospitalmis.entity.ImagingProcedure;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImagingProcedureRepository extends JpaRepository<ImagingProcedure, Integer>{ }
+public interface ImagingProcedureRepository extends JpaRepository<ImagingProcedure, Long>{ }
